moved view model class to trello_items
